[BUGFIX] Do not try to download files not present in sys_file
Since the extension hooks into every file exists call of TYPO3 FAL
it happens that it tries to download files from a remote,
when a noew file is put into a storage (e.g. during upload).

To prevent that a check is performed whether a file object
of the current file in question exists and if so, dowload
is omitted.